+++ include/ocudu/ran/precoding/precoding_matrix_indicator.h
@@ -23,7 +23,7 @@ namespace ocudu {
 /// Precoding Matrix Indicator (PMI) for two antenna ports.
 struct pmi_two_antenna_port {
   /// PMI codebook index from TS38.214 Table 5.2.2.2.1-1
-  unsigned pmi;
+  uint8_t pmi;
 };
 
 /// \brief Precoding Matrix Indicator (PMI) field for Type I Single-Panel codebook.
@@ -32,14 +32,14 @@ struct pmi_two_antenna_port {
 struct pmi_typeI_single_panel {
   /// Single-panel topology configuration.
   pmi_codebook_typeI_single_panel panel_config;
-  /// PMI parameter \f$i_{1,1}\f$.
-  unsigned i_1_1;
+  /// PMI parameter \f$i_{1,1}\f$. Values: {0,...,N_1*O_1 - 1}, at most {0,...,63}.
+  uint8_t i_1_1;
   /// PMI parameter \f$i_{1,2}\f$. Only available for \f$N_2 > 1\f$ or \f$\upsilon > 2\f$.
-  std::optional<unsigned> i_1_2;
+  std::optional<uint8_t> i_1_2;
   /// PMI parameter \f$i_{1,3}\f$. Only available for \f$\upsilon \in \{2,3,4\}\f$.
-  std::optional<unsigned> i_1_3;
+  std::optional<uint8_t> i_1_3;
   /// PMI parameter \f$i_2\f$.
-  unsigned i_2;
+  uint8_t i_2;
 };
 
 /// \brief Precoding Matrix Indicator (PMI) field for the Type II codebook.
@@ -54,7 +54,7 @@ struct pmi_typeII {
   /// Each coefficient vector holds one entry per beam and polarization, i.e. 2*L entries.
   struct layer_coefficients {
     /// PMI parameter \f$i_{1,3,l}\f$. Strongest-coefficient index, {0, ..., 2L-1}.
-    unsigned i_1_3;
+    uint8_t i_1_3;
     /// PMI parameter \f$i_{1,4,l}\f$. Wideband amplitude indices, one per beam and polarization.
     static_vector<uint8_t, max_nof_typeII_beams * 2> i_1_4;
     /// PMI parameter \f$i_{2,1,l}\f$. Phase coefficient indices, one per beam and polarization.
@@ -66,11 +66,12 @@ struct pmi_typeII {
 
   /// Type II codebook configuration.
   pmi_codebook_typeII config;
-  /// PMI parameter \f$i_{1,1}\f$, encoding the beam selection parameters \f$(q_1, q_2)\f$ inside a group.
-  unsigned i_1_1;
+  /// PMI parameter \f$i_{1,1}\f$, encoding the beam selection parameters \f$(q_1, q_2)\f$ inside a group. Values:
+  /// {0,...,O_1*O_2 - 1}, at most {0,...,15}.
+  uint8_t i_1_1;
   /// PMI parameter \f$i_{1,2}\f$, the combinatorial \f$L\f$-beam group selection index, as per TS38.214
-  /// Table 5.2.2.2.3-1.
-  unsigned i_1_2;
+  /// Table 5.2.2.2.3-1. Values: {0,...,C(N_1*N_2, L) - 1}, at most {0,...,1819}.
+  uint16_t i_1_2;
   /// Per-layer combining coefficients, one entry per layer. The Type II codebook supports rank 1 or 2.
   static_vector<layer_coefficients, 2> layers;
 };
